+++ features/contacts_ui.py
@@ -3,7 +3,7 @@ from tkinter import ttk, messagebox, simpledialog, filedialog, Toplevel, Listbox
 import sqlite3
 from datetime import datetime
 import pandas as pd
-from contact_dialog import AddContactDialog
+from .contact_dialog import AddContactDialog
 from .common import DB_FILE, load_column_widths, save_column_widths, get_settings, get_all_group_names
 
 
@@ -509,7 +509,7 @@ def send_emails_dialog(tree):
             status_var.set("Amazon SES credentials missing in Settings.")
             send_btn.config(state=tk.NORMAL)
             return
-        import email_utils
+        from services import email_utils
         def send_thread():
             history_conn = sqlite3.connect(DB_FILE)
             hc = history_conn.cursor()
@@ -559,7 +559,7 @@ def send_emails_dialog(tree):
 
 def import_contacts_dialog(tree):
     from tkinter import filedialog, messagebox, simpledialog, Toplevel, Listbox, MULTIPLE, Button, Label, END
-    from contacts import import_contacts_from_csv
+    from services.contacts import import_contacts_from_csv
     import sqlite3
     # Step 1: File selection
     filename = filedialog.askopenfilename(
