+++ features/email.py
@@ -4,7 +4,7 @@ import sqlite3
 import time
 import threading
 from datetime import datetime
-import email_utils
+from services import email_utils
 from .common import DB_FILE, get_settings, center_window, get_all_group_names
